REFACTOR test_transform suites to single test module
